Split up Brig.User.Search.Index into seperate modules.
